+[IGAND-676] display titles and added lazy loading on user page

+++ app/src/main/java/igraal/com/poc_deezer_vincent/adapter/AdapterIdCallBack.java
@@ -5,5 +5,5 @@ package igraal.com.poc_deezer_vincent.adapter;
  */
 
 public interface AdapterIdCallBack {
-    public void onCallBack(int id);
+    public void onCallBack(long id);
 }
